==> data_analysis/1 train test split.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split

path = '/Users/tristangarcia/desktop/hp-pred/data/wa/'
data = pd.read_csv(f'{path}wa_data.csv')

data.columns

# Renaming propertyTypes
data.rename(columns={'propertyTypes':'propertyType'}, inplace=True)

# Removing irrelevant columns. Alot of these columns will be re-added by feature engineering. 
data.drop(columns=['listingId','unitNumber','state','agency','promoLevel','soldMonth','daysListed','inspectionsCount','isRural',
                   'hasDescription','hasFloorplan','hasDisplayPrice','hasPhoto','photoCount','suburb_medianPrice','suburb_medianRentPrice',
                   'suburb_entryLevelPrice','suburb_luxuryLevelPrice','primary','primaryDistance','primaryType',
                   'secondary','secondaryDistance','secondaryType','listingUrl'], inplace=True)

# SPLIT

# train test split
train,test = train_test_split(data, test_size=0.3, random_state=222, shuffle=True)
# validation split
validation,test = train_test_split(test, test_size=0.5, random_state=222, shuffle=True)

# Train

train.shape

train.head()

# Validation

validation.shape

validation.head()

# Test

test.shape

test.head()

# Writing to csv

train.to_csv(f'{path}wa_train.csv', index=False)
validation.to_csv(f'{path}wa_validation.csv', index=False)
test.to_csv(f'{path}wa_test.csv', index=False)

train.columns
