==> data_analysis/2 cleaning.ipynb <==
import pandas as pd
pd.set_option('display.max_columns', None)
pd.set_option("max_colwidth", None)
pd.set_option("max_seq_items", None)
pd.set_option('display.float_format', '{:.2f}'.format)

import certifi
import geopy.geocoders
from geopy.geocoders import Nominatim
import ssl
import time

path = '/Users/tristangarcia/desktop/hp-pred/data/wa/'

train = pd.read_csv(f'{path}wa_train.csv')
test = pd.read_csv(f'{path}wa_validation.csv')  # validation set

print(train.shape)
print(test.shape)

# Filter

train['soldYear'].value_counts()

# Not enough data for 2017 and earlier
train = train[train['soldYear'] > 2017]
test = test[test['soldYear'] > 2017]

### Sold price

# Removing unrealistic house prices
train = train[(train['soldPrice'] > 100000)]
train = train[train['soldPrice'] < 10000000]

test = test[(test['soldPrice'] > 100000)]
test = test[(test['soldPrice'] < 10000000)]

na_cols = ['suburb','bathrooms','bedrooms','parking',
           'soldYear','soldPrice']
# Removing missing rows for required variables soldPrice
train.dropna(subset=na_cols, inplace=True)
test.dropna(subset=na_cols, inplace=True)

print(train.shape)
print(test.shape)

# Reformatting 

train.shape

### Property type

train['propertyType'].value_counts()

### Only possible entries:
- House
- Apartment
- townhouse??


train['propertyType'] = train['propertyType'].str.split(',')
test['propertyType'] = test['propertyType'].str.split(',')

def reformat_propertyTypes(row):
    # Valid property types
    types = ['House','Apartment']
    # This deals with NaN values
    if not isinstance(row, list):
        return 'Unknown'
    # Go through all property types
    for t in row:
        if t in types:
            # Return the first instance of a valid property type
            return t
    # Return 'Unkown' if no valid property types found
    return 'Unknown'
        
# Applying the function
train['propertyType'] = train['propertyType'].apply(reformat_propertyTypes)
test['propertyType'] = test['propertyType'].apply(reformat_propertyTypes)

# Changing the rural values from propertyTypes to House
train['propertyType'] = train['propertyType'].apply(lambda x: 'unit' if x.lower() =='apartment' else x)
test['propertyType'] = test['propertyType'].apply(lambda x: 'unit' if x.lower() =='apartment' else x)

# Removing unknown propertyTypes
train = train[(train['propertyType'] != 'Unknown')]
test = test[(test['propertyType'] != 'Unknown')]

train['propertyType'].value_counts()

train.shape

### Features

train['features'].value_counts()

train['features'][3][0]

We can see that each instance of the variables 'features' looks like a list but is actually a string

# Converting the 'string lists' into lists
# https://stackoverflow.com/questions/1894269/how-to-convert-string-representation-of-list-to-a-list
import ast

train['features'] = train['features'].apply(lambda x: ast.literal_eval(x) if pd.notnull(x) else [])
test['features'] = test['features'].apply(lambda x: ast.literal_eval(x) if pd.notnull(x) else [])

### Lowercasing categorical variables

train = train.map(lambda x: x.lower() if isinstance(x, str) else x)
test = test.map(lambda x: x.lower() if isinstance(x,str) else x)

# Lowercasing features variable
def lower_features(feature_list):
    f = []
    if len(feature_list) == 0:
        return feature_list
    else:
        for feature in feature_list:
            f.append(feature.lower())
    return f

train['features'] = train['features'].apply(lambda x: lower_features(x))
test['features'] = test['features'].apply(lambda x: lower_features(x))

# Missing values

train.isnull().sum()

### Latitude and Longitude

We will be applying two methods in order to fill latitude and longitude
- Method 1 will be to use Nominatim to search for the address
- Method 2 will be to use a csv we made to merge and fill using the latitude and longitude of the suburb

# Initialise nominatim
ctx = ssl.create_default_context(cafile=certifi.where())
geopy.geocoders.options.default_ssl_context = ctx
nom = Nominatim(user_agent="tristan_scrape")


def format_address(house):
    parts = []
    # Check for street and street number, append if both are present
    if pd.notnull(house['street']) and pd.notnull(house['streetNumber']):
        parts.append(f'{house["streetNumber"]} {house["street"]}')
    elif pd.notnull(house['street']):
        parts.append(house['street'])
    
    # Check and append if suburb and postcode are present
    if pd.notnull(house['suburb']):
        parts.append(house['suburb'])
    if pd.notnull(house['postcode']):
        parts.append(f'{house["postcode"].astype(int)}')
    if parts:  # if there's at least one part present, append the state
        parts.append('Western Australia')
    # Join the parts with a comma and return the result
    print(parts)
    return ', '.join(parts) if parts else 'None'

def get_coords(address):
    # Searches Nominatim of the address
    location = nom.geocode(address)
    time.sleep(1)
    if location:
        return location.latitude, location.longitude
    return None, None


def coords_fill(df):
    # Getting index of data to fill
    index = df.loc[(df['latitude'].isnull()) & (df['suburb'].notnull())].index
    for idx in index:
        house = df.loc[idx]
        address = format_address(house)
        latitude, longitude = get_coords(address)
        # Assigning new latitude and longitude
        df.loc[idx, 'latitude'] = latitude
        df.loc[idx, 'longitude'] = longitude
    return df

train = coords_fill(train)
test = coords_fill(test)

suburbs_path = '/Users/tristangarcia/desktop/hp-pred/data/'
suburb_statistics = pd.read_csv(f'{suburbs_path}wa_suburb_statistics.csv')

suburb_statistics.head()

coord_df = suburb_statistics[['suburb','latitude','longitude']]
print(coord_df.shape)
coord_df.head()

# Merging train with coord_df
train = pd.merge(train, coord_df, on=['suburb'], how='left', suffixes=('', '_from_coord'))
test = pd.merge(test, coord_df, on=['suburb'], how='left', suffixes=('', '_from_coord'))

# Filling missing Latitude and Longitude
train['latitude'] = train['latitude'].fillna(train['latitude_from_coord'])
train['longitude'] = train['longitude'].fillna(train['longitude_from_coord'])
test['latitude'] = test['latitude'].fillna(test['latitude_from_coord'])
test['longitude'] = test['longitude'].fillna(test['longitude_from_coord'])

# Dropping the extra columns from coord_df
train.drop(['latitude_from_coord', 'longitude_from_coord'], axis=1, inplace=True)
test.drop(['latitude_from_coord', 'longitude_from_coord'], axis=1, inplace=True)

train.isnull().sum()

# KNN Imputation

from sklearn.neighbors import KNeighborsRegressor

### Land area
We will be using latitude and longitude to impute on landArea. We suspect that there is a difference in landArea for rural and non-rural areas. There also seems to be less recorded neighbours for rural areas. To account for this, the data will be split into rural and non-rural areas before imputation. Additionally, rural areas will find only 1 nearest neighbour will non-rural will find 3 nearest neighbours

train.reset_index(drop=True,inplace=True)
test.reset_index(drop=True,inplace=True)
# Seperating null and non-null 
x = train.loc[train['landArea'].notnull(), ['latitude','longitude','landArea']]
y = train.loc[train['landArea'].isnull(), ['latitude','longitude','landArea']]
y_test = test.loc[test['landArea'].isnull(), ['latitude','longitude','landArea']]

print(x.shape)
print(y.shape)

imputer = KNeighborsRegressor(n_neighbors=1)
# Fitting on the training set
imputer.fit(x[['latitude','longitude']], x['landArea'])
# Imputing on missing landArea 
predict = imputer.predict(y[['latitude', 'longitude']])
train.loc[y.index,'landArea'] = predict
# Imputing on missing landArea on validation set using training data
predict_test = imputer.predict(y_test[['latitude', 'longitude']])
test.loc[y_test.index,'landArea'] = predict_test

# Writing to file

print(train.shape)
print(test.shape)

train.to_pickle(f'{path}wa_train_clean.pkl')
test.to_pickle(f'{path}wa_validation_clean.pkl')

train.head(20)

train.isnull().sum()
